added backend form validation
